Made Injector#setParentInjector public to enable turning independently created injectors into child injectors.
This is mainly meant to be used in the context of Robotlegs, which subclasses Injector to allow for using swappable injector implementations through a uniform interface.
Because SwiftSuspenders must not have a dependency on Robotlegs, we can't internally create an instance of Robotlegs' SwiftSuspendersInjector. Using the public setParentInjector method, it's possible to manually instantiate SwiftSuspendersInjector and set the parent context's injector as the parent injector afterwards.