Completely implemented and hooked up a reflector based on describeTypeJSON from the avmplus package
As describeTypeJSON might go away or change in the future, the implementation is completely optional and will automatically be used if availability is detected at runtime. Otherwise, the old describeType-based reflector will be used.

Additionally, some small fixes and cleanups are contained in this patch: They were pretty hard to un-entangle, so I couldn't move them to another commit.