Reduced injector overhead by about 30% by introducing several small-ish optimizations
Specifically, mappings are now stored and ( in relevant places) queried by class name in addition to class identity to remove the need to do expensive class-lookups all the time. Additionally, the logic for applying lists of injection points was extracted from injectInto to remove the need to re-lookup the type for an instance in case of the injector doing the instantiation itself.